Figured out to change feedforward behavior
Previously we called matcaffe_init to initialize the network, and in the
function, the network will not reinitialize if it already has been initialized.
So changes to the network will not be reflected (e.g., even if you `clear all`
on MATLAB).

Now we call caffe() directly from our own script to reinitialize the network
every time it is run.