Able to finetune network with our solver
Previously the solver/train_val scripts couldn't be run because they weren't in
the CAFFENET installation directory. Now I've tweaked the paths to be absolute
to my directory so it can. If you wanna run them you'll have to change the
paths in the solver to your own directory.

caffe train is expected to be run from CAFFENET installation directory, e.g.
./build/tools/caffe train -solver
~/captioned-image-retrieval/caffenet/finetune_solver.prototxt -weights
~/captioned-image-retrieval/caffenet/bvlc_reference_caffenet.caffemodel