Prediction script now predicts train/dev/test
Previously we just predicted on all images in an image dir, which is iffy
because there's 91 images (out of 8091) that aren't in train/dev/test. Now we
take the list of image filenames from the file listings provided by the
Flickr8k dataset, so numbers round of to nice number, e.g. divisible by 10.